2024/09/23 リダイレクト先をreportsに変更

+++ src/main/java/com/techacademy/repository/ReportRepository.java
@@ -3,5 +3,5 @@ package com.techacademy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.techacademy.entity.Report;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
 }
