+++ src/main/java/com/techacademy/service/ReportService.java
@@ -4,11 +4,9 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.techacademy.entity.Report;
-import com.techacademy.repository.EmployeeRepository;
 import com.techacademy.repository.ReportRepository;
 
 @Service
@@ -27,7 +25,7 @@ public class ReportService {
     }
 
     // 1件を検索
-    public Report findByCode(Long id) {
+    public Report findByCode(Integer id) {
         // findByIdで検索
         Optional<Report> option = reportRepository.findById(id);
         // 取得できなかった場合はnullを返す
